requirements.txt
accelerate
addict
annotated-doc
annotated-types
antlr4-python3-runtime
anyio
argcomplete
attrs
av
certifi
cfgv
charset-normalizer
click
colorama
contourpy
cycler
decorator
decord
diffusers
distlib
easydict
einops
evo
fastapi
filelock
fonttools
fsspec
ftfy
gitdb
GitPython
h11
hf-xet
hf_transfer
httpcore
httpx
huggingface-hub
identify
idna
ImageIO
imageio-ffmpeg
importlib_metadata
iopath
Jinja2
kiwisolver
lmdb
lz4
markdown-it-py
MarkupSafe
matplotlib
mdurl
moviepy
mpmath
natsort
networkx
ninja
nodeenv
numexpr
numpy
nvitop
omegaconf
opencv-python
opt-einsum-fx
opt_einsum
packaging
pandas
pillow
pillow_heif
platformdirs
plyfile
portalocker
pre_commit
proglog
protobuf
psutil
pyarrow
pycocotools
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
PyYAML
regex
requests
rerun-sdk
rich
rosbags
ruamel.yaml
safetensors
scipy
seaborn
sentry-sdk
setuptools
shellingham
six
smmap
starlette
sympy
timm
tokenizers
tqdm
transformers
trimesh
typer
typing-inspection
typing_extensions
urllib3
uvicorn
virtualenv
wandb
wcwidth
wheel
zipp
zstandard
peft